Handle account route with proper ACID transaction

+++ src/lib/auth/accountDeletion.ts
@@ -5,42 +5,31 @@ import mongoosePromise, {
 } from '@/lib/database/mongoose';
 
 export const userDeleter = async (id: string) => {
-	try {
-		await mongoosePromise();
-		await Promise.all([
-			new Promise(async (resolve, reject): Promise<void> => {
-				try {
-					await Account.findOneAndDelete({ userId: id })
-						.where('provider')
-						.equals('spotify')
-						.exec();
-					resolve(null);
-				} catch {
-					const message = 'Error deleting account';
-					reject({ message });
-				}
-			}),
-			new Promise(async (resolve, reject): Promise<void> => {
-				try {
-					await User.findByIdAndDelete(id);
-					resolve(null);
-				} catch {
-					const message = 'Error deleting user';
-					reject({ message });
-				}
-			}),
-			new Promise(async (resolve, _): Promise<void> => {
-				try {
-					await sessionDeleter(id);
-				} catch {
-					// Silent catch for sessions: nbd
-				}
-				resolve(null);
-			}),
-		]);
-	} catch (e: any) {
-		throw { message: e.message || 'Unknown error deleting user' };
+	const db = await mongoosePromise();
+	if (!db) throw new Error();
+	const session = await db.startSession();
+	let onceFlag = true;
+	while (onceFlag) {
+		try {
+			session.startTransaction();
+			await Promise.all([
+				Account.findOneAndDelete({ userId: id })
+					.where('provider')
+					.equals('spotify')
+					.exec(),
+				User.findByIdAndDelete(id).exec(),
+			]);
+			session.commitTransaction();
+		} catch {
+			session.abortTransaction();
+			if (onceFlag) {
+				onceFlag = false;
+				continue;
+			}
+			throw new Error();
+		}
 	}
+	session.endSession();
 	return;
 };
